refactor: move init code from main to instrumentor package
